# [inductor] optimize lowering for empty-related operators