--- a/torch/_inductor/lowering.py
+++ b/torch/_inductor/lowering.py
@@ -1628,32 +1628,18 @@ def empty(
     *size,
     names=None,
     dtype=None,
-    layout=0,
+    layout=None,
     device=None,
     pin_memory=None,
     memory_format=None,
 ):
     assert names is None
-    assert not pin_memory
-    assert layout in (0, torch.strided)
     assert memory_format in (None, torch.contiguous_format)
-    device = decode_device(device)
-    dtype = dtype or torch.get_default_dtype()
     if len(size) == 1 and isinstance(size[0], (list, tuple, torch.Size)):
-        size = tuple(size[0])
-    size = [sympy.expand(s) for s in size]
-    pointwise = _full(0, device, dtype, size)
-    pointwise.realize()
-    buffer = pointwise.data.data
-    buffer.data.ranges = [0] * len(size)
-    assert isinstance(buffer, ir.ComputedBuffer)
-    buffer.layout = ir.FixedLayout(
-        device=device,
-        dtype=dtype,
-        size=[sympy.expand(s) for s in size],
-        stride=ir.FlexibleLayout.contiguous_strides(size),
+        size = list(size[0])
+    return empty_strided(
+        size, None, dtype=dtype, layout=layout, device=device, pin_memory=pin_memory
     )
-    return pointwise
 
 
 def create_tensor_like(creation_fn):
@@ -1713,24 +1699,13 @@ register_lowering(aten.new_ones)(new_constant(1))
 
 @register_lowering(aten.new_empty)
 def new_empty(x, size, *, dtype=None, layout=None, device=None, pin_memory=None):
-    assert isinstance(size, (list, type))
-    assert not pin_memory
-    assert not layout or layout == torch.strided
-    dtype = decode_dtype(dtype) or x.get_dtype()
-    device = device or x.get_device()
-    size = [sympy.Integer(s) for s in size]
-    pointwise = _full(0, device, dtype, size)
-    pointwise.realize()
-    buffer = pointwise.data.data
-    buffer.data.ranges = [0] * len(size)
-    assert isinstance(buffer, ir.ComputedBuffer)
-    buffer.layout = ir.FixedLayout(
-        device=device,
-        dtype=dtype,
-        size=[sympy.expand(s) for s in size],
-        stride=ir.FlexibleLayout.contiguous_strides(size),
+    if dtype is None:
+        dtype = x.get_dtype()
+    if device is None:
+        device = x.get_device()
+    return empty_strided(
+        size, None, dtype=dtype, layout=layout, device=device, pin_memory=pin_memory
     )
-    return pointwise
 
 
 @register_lowering(aten.empty_strided)
@@ -1738,7 +1713,7 @@ def empty_strided(
     size, stride, *, dtype=None, layout=None, device=None, pin_memory=None
 ):
     assert isinstance(size, (list, type))
-    assert isinstance(stride, (list, type))
+    assert isinstance(stride, (list, type)) or stride is None
     assert not pin_memory
     assert not layout or layout == torch.strided
     dtype = decode_dtype(dtype) or torch.get_default_dtype()
@@ -1746,13 +1721,20 @@ def empty_strided(
     pointwise = _full(fill_value=0, device=device, dtype=dtype, size=size)
     pointwise.realize()
     buffer = pointwise.data.data
+    # explicitly set ranges to zeros in order to make a NopKernelSchedulerNode
     buffer.data.ranges = [0] * len(size)
     assert isinstance(buffer, ir.ComputedBuffer)
+    size = [sympy.expand(s) for s in size]
+    stride = (
+        [sympy.expand(s) for s in stride]
+        if stride
+        else ir.FlexibleLayout.contiguous_strides(size)
+    )
     buffer.layout = ir.FixedLayout(
         device=device,
         dtype=dtype,
-        size=[sympy.expand(s) for s in size],
-        stride=[sympy.expand(s) for s in stride],
+        size=size,
+        stride=stride,
     )
     return pointwise