@ngimel 2023-01-03
tbh, I don't see a big problem with current lowering - yeah, for a single `empty` call in the graph it'll generate unneeded kernel, but we don't have graphs with a single `empty` call, and if something else is happening it's getting fused.